% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdas-package.R
\docType{package}
\name{isdas-package}
\alias{isdas}
\alias{isdas-package}
\title{isdas: Companion Package for Book An Introduction to Spatial Data Analysis and Statistics}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/isdas}
  \item \url{https://paezha.github.io/isdas/}
  \item Report bugs at \url{https://github.com/paezha/isdas/issues}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

}
\keyword{internal}
