% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdas-package.R
\docType{package}
\name{isdas}
\alias{isdas}
\title{isdas}
\description{
A package with data sets and functions used in the book An Introduction to Spatial Data Analysis and Statistics: A Course in \code{R}.
}
\author{
Antonio Paez \link{cre, aut}, \email{paezha@mcmaster.ca}, Anastassios Dardas \link{ctb}, Raj Ubhi \link{ctb}
}
