% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpointmean.R
\name{kpointmean}
\alias{kpointmean}
\title{A function that implements k-point mean interpolation}
\usage{
kpointmean(source_xy, target_xy, z, k, longlat = FALSE)
}
\arguments{
\item{source_xy}{A simple features object of type POINT with the source points for interpolation}

\item{target_xy}{A simple features object of type POINT with the target points for interpolation}

\item{z}{The name of the column with the variable in \code{source_xy} to interpolate}

\item{k}{A number that defines the number of neighbors used for interpolation}

\item{longlat}{TRUE if coordinates are in long/lat}
}
\value{
A simple features object.
}
\description{
This function calculates the spatial mean based on the k-nearest neighbors of a focal point or set of focal points.
}
\keyword{spatial}
