% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdas-package.R
\docType{package}
\name{isdas-package}
\alias{isdas}
\alias{isdas-package}
\title{isdas: Companion Package for Book An Introduction to Spatial Data Analysis and Statistics}
\description{
This package contains data sets, custom functions, and document templates to work with the book An Introduction to Spatial Data Analysis and Statistics: A Course in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/isdas}
  \item \url{https://paezha.github.io/isdas/ https://paezha.github.io/spatial-analysis-with-r/}
  \item \url{https://paezha.github.io/isdas/}
  \item Report bugs at \url{https://github.com/paezha/isdas/issues}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Rajveer Ubhi \email{ubhir20@gmail.com}
  \item Anastassios Dardas
}

}
\keyword{internal}
